# docs/swagger.yaml
definitions:
  domain.Loan:
    properties:
      amount:
        type: number
      created_at:
        type: string
      duration:
        type: integer
      id:
        type: string
      interest:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.Log:
    properties:
      activity:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
  domain.User:
    properties:
      bio:
        type: string
      contact:
        type: string
      email:
        type: string
      id:
        type: string
      imageuri:
        type: string
      isadmin:
        type: boolean
      isverified:
        type: boolean
      joinedat:
        type: string
      password:
        type: string
      refreshtoken:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: dagmmesfin99@gmail.com
    name: Dagim Mesfin
  description: Loan Tracker API for managing loans and user accounts.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Loan Tracker API
  version: "1.0"
paths:
  /admin/loans:
    get:
      consumes:
      - application/json
      description: View all loans
      parameters:
      - description: Page number
        in: query
        name: pgnum
        type: integer
      - description: Loan status
        in: query
        name: status
        type: string
      - description: Order by
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Loan'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View all loans
      tags:
      - Admin
  /admin/loans/{loan_id}:
    delete:
      consumes:
      - application/json
      description: Delete a loan
      parameters:
      - description: Loan ID
        in: path
        name: loan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a loan
      tags:
      - Admin
  /admin/loans/{loan_id}/status:
    patch:
      consumes:
      - application/json
      description: Approve or reject a loan
      parameters:
      - description: Loan ID
        in: path
        name: loan_id
        required: true
        type: string
      - description: Loan status (approved/rejected)
        in: body
        name: status
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Approve or reject a loan
      tags:
      - Admin
  /admin/logs:
    get:
      consumes:
      - application/json
      description: View logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Log'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View logs
      tags:
      - Admin
  /admin/user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - User
  /admin/users:
    get:
      consumes:
      - application/json
      description: View all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: View all users
      tags:
      - User
  /loan/{loan_id}:
    get:
      consumes:
      - application/json
      description: Get loan details
      parameters:
      - description: Loan ID
        in: path
        name: loan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Loan'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get loan details
      tags:
      - Loan
  /loan/apply:
    post:
      consumes:
      - application/json
      description: Apply for a loan
      parameters:
      - description: Loan details
        in: body
        name: loan
        required: true
        schema:
          $ref: '#/definitions/domain.Loan'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Loan'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Apply for a loan
      tags:
      - Loan
  /user/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - User
  /user/logout:
    get:
      consumes:
      - application/json
      description: Logout user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - User
  /user/password-reset:
    post:
      consumes:
      - application/json
      description: Forgot password
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password
      tags:
      - User
  /user/password-update:
    post:
      consumes:
      - application/json
      description: Reset password
      parameters:
      - description: Reset token
        in: query
        name: token
        required: true
        type: string
      - description: New password
        in: body
        name: new_password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      description: Get user profile
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - User
  /user/token-refresh:
    get:
      consumes:
      - application/json
      description: Refresh user token
      parameters:
      - description: Refresh token
        in: query
        name: refresh-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh user token
      tags:
      - User
  /user/update:
    put:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user details
      tags:
      - User
  /user/verify-email:
    get:
      consumes:
      - application/json
      description: Verify user email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify user email
      tags:
      - User
swagger: "2.0"